+++ tests/testthat/test-marginalmeans.R
@@ -1,5 +1,7 @@
 skip_if_not_installed("emmeans")
 skip_if_not_installed("broom")
+requiet("emmeans")
+requiet("broom")
 
 dat <- mtcars
 dat$am <- as.logical(dat$am)
@@ -17,9 +19,9 @@ test_that("changing the prediction grid changes marginal means", {
     mm1 <- marginalmeans(mod, variables = "cyl")
     mm2 <- marginalmeans(mod, variables = "cyl", variables_grid = "vs")
     mm3 <- marginalmeans(mod, variables = "cyl", variables_grid = "am")
-    expect_false(all(mm1$predicted == mm2$predicted))
-    expect_false(all(mm1$predicted == mm3$predicted))
-    expect_false(all(mm2$predicted == mm3$predicted))
+    expect_false(all(mm1$marginalmean == mm2$marginalmean))
+    expect_false(all(mm1$marginalmean == mm3$marginalmean))
+    expect_false(all(mm2$marginalmean == mm3$marginalmean))
 })
 
 test_that("tidy and glance", {
@@ -31,26 +33,33 @@ test_that("tidy and glance", {
     expect_equal(dim(ti), c(5, 8))
 })
 
-test_that("marginalmeans variance: link scale or linear", {
+test_that("marginalmeans vs. emmeans: poisson link or response", {
     dat <- mtcars
     dat$am <- factor(dat$am)
     dat$cyl <- factor(dat$cyl)
     mod <- glm(gear ~ cyl + am, data = dat, family = poisson)
-    tmp <- marginalmeans(mod)
-    expect_false("std.error" %in% colnames(tmp))
-    tmp <- marginalmeans(mod, type = "link")
-    expect_true("std.error" %in% colnames(tmp))
+    # link
+    mm <- tidy(marginalmeans(mod, variables = "cyl", type = "link"))
+    em <- tidy(emmeans(mod, specs = "cyl"))
+    expect_equal(mm$estimate, em$estimate)
+    expect_equal(mm$std.error, em$std.error)
+    # response
+    mm <- tidy(marginalmeans(mod, variables = "cyl", type = "response"))
+    em <- tidy(emmeans(mod, specs = "cyl", transform = "response"))
+    expect_equal(mm$estimate, em$rate)
+    expect_equal(mm$std.error, em$std.error, tolerance = .0001)
 })
+
 
 test_that("simple marginal means", {
     mod <- lm(mpg ~ cyl + am + hp, dat)
     em <- broom::tidy(emmeans::emmeans(mod, "cyl"))
     me <- marginalmeans(mod, variables = "cyl")
-    expect_equal(me$predicted, em$estimate)
+    expect_equal(me$marginalmean, em$estimate)
     expect_equal(me$std.error, em$std.error)
     em <- broom::tidy(emmeans::emmeans(mod, "am"))
     me <- marginalmeans(mod, variables = "am")
-    expect_equal(me$predicted, em$estimate)
+    expect_equal(me$marginalmean, em$estimate)
     expect_equal(me$std.error, em$std.error)
 })
 
@@ -60,11 +69,11 @@ test_that("interactions", {
     em <- suppressMessages(broom::tidy(emmeans::emmeans(mod, "cyl")))
     me <- suppressWarnings(marginalmeans(mod, variables = "cyl"))
     expect_warning(marginalmeans(mod, variables = "cyl"), regex = "interactions")
-    expect_equal(me$predicted, em$estimate)
+    expect_equal(me$marginalmean, em$estimate)
     em <- suppressMessages(broom::tidy(emmeans::emmeans(mod, "am")))
     me <- suppressWarnings(marginalmeans(mod, variables = "am"))
     expect_warning(marginalmeans(mod, variables = "am"), regex = "interactions")
-    expect_equal(me$predicted, em$estimate)
+    expect_equal(me$marginalmean, em$estimate)
 })
 
 test_that("error: no factor", {